+++ SwapIt.BL/DTOs/ServiceFilterDto.cs
@@ -9,7 +9,7 @@ namespace SwapIt.BL.DTOs
     public class ServiceFilterDto
     {
         public string ServiceName { get; set; }
-        public float ServicePrice { get; set; }
+        public int ServicePrice { get; set; }
         public int ServiceProviderId { get; set; }
         public int CategoryId { get; set; }
